% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZoopCounter.R
\name{ZoopCounter}
\alias{ZoopCounter}
\title{Open the Zooplankton Counter}
\usage{
ZoopCounter(data, shortcuts = NULL)
}
\arguments{
\item{data}{A vector of names}

\item{shortcuts}{A vector of characters representing the shortcut keys used to add to a species' count. For example, c("Z", "X", "C"...). Has a default setting if none is specified.}
}
\value{
A Zooplankton Counter instance
}
\description{
Opens up the Zooplankton counting tool
}
